refactor: extract shared TodoMVC fixture to tests/package-common
Move common test schema, tables, events, and store creation logic
to a shared package that can be used by both React and Solid tests.
This reduces duplication and ensures consistency across framework tests.
