+++ packages/@livestore/react/package.json
@@ -16,6 +16,7 @@
   "devDependencies": {
     "@livestore/adapter-web": "workspace:*",
     "@livestore/utils-dev": "workspace:*",
+    "@local/tests-package-common": "workspace:*",
     "@opentelemetry/sdk-trace-base": "catalog:",
     "@testing-library/dom": "^10.4.1",
     "@testing-library/react": "^16.3.0",
