+++ packages/@livestore/solid/package.json
@@ -17,6 +17,7 @@
     "@effect/vitest": "catalog:",
     "@livestore/adapter-web": "workspace:*",
     "@livestore/utils-dev": "workspace:*",
+    "@local/tests-package-common": "workspace:*",
     "@opentelemetry/sdk-trace-base": "catalog:",
     "@solidjs/testing-library": "^0.8.10",
     "@testing-library/jest-dom": "^6.9.1",
