+++ packages/@livestore/solid/src/__tests__/fixture.tsx
@@ -1,127 +1,40 @@
-import { makeInMemoryAdapter } from '@livestore/adapter-web'
-import { provideOtel, type UnknownError } from '@livestore/common'
-import { Events, makeSchema, State } from '@livestore/common/schema'
-import type { LiveStoreSchema, SqliteDsl, Store } from '@livestore/livestore'
-import { createStore, StoreInternalsSymbol } from '@livestore/livestore'
-import { omitUndefineds } from '@livestore/utils'
-import { Effect, Schema, type Scope } from '@livestore/utils/effect'
-import type * as otel from '@opentelemetry/api'
-import type { JSX } from 'solid-js'
-
+import type { UnknownError } from '@livestore/common'
+import type { Store } from '@livestore/livestore'
+import { StoreInternalsSymbol } from '@livestore/livestore'
+import { Effect, type Scope } from '@livestore/utils/effect'
+import {
+  type AppState,
+  type CreateTodoMvcStoreOptions,
+  createTodoMvcStore,
+  events,
+  type Filter,
+  schema,
+  type Todo,
+  tables,
+} from '@local/tests-package-common/todomvc-fixture'
+import type * as Solid from 'solid-js'
 import * as LiveStoreSolid from '../mod.ts'
 
-export { StoreInternalsSymbol }
+// Re-export shared types, schema, and StoreInternalsSymbol for tests
+export { events, schema, StoreInternalsSymbol, tables }
+export type { AppState, Filter, Todo }
 
-export type Todo = {
-  id: string
-  text: string
-  completed: boolean
-}
-
-export type Filter = 'all' | 'active' | 'completed'
-
-export type AppState = {
-  newTodoText: string
-  filter: Filter
-}
-
-const todos = State.SQLite.table({
-  name: 'todos',
-  columns: {
-    id: State.SQLite.text({ primaryKey: true }),
-    text: State.SQLite.text({ default: '', nullable: false }),
-    completed: State.SQLite.boolean({ default: false, nullable: false }),
-  },
-})
-
-const app = State.SQLite.table({
-  name: 'app',
-  columns: {
-    id: State.SQLite.text({ primaryKey: true, default: 'static' }),
-    newTodoText: State.SQLite.text({ default: '', nullable: true }),
-    filter: State.SQLite.text({ default: 'all', nullable: false }),
-  },
-})
-
-const userInfo = State.SQLite.clientDocument({
-  name: 'UserInfo',
-  schema: Schema.Struct({
-    username: Schema.String,
-    text: Schema.String,
-  }),
-  default: { value: { username: '', text: '' } },
-})
-
-const AppRouterSchema = State.SQLite.clientDocument({
-  name: 'AppRouter',
-  schema: Schema.Struct({
-    currentTaskId: Schema.String.pipe(Schema.NullOr),
-  }),
-  default: {
-    value: { currentTaskId: null },
-    id: 'singleton',
-  },
-})
-
-const kv = State.SQLite.clientDocument({
-  name: 'Kv',
-  schema: Schema.Any,
-  default: { value: null },
-})
-
-export const events = {
-  todoCreated: Events.synced({
-    name: 'todoCreated',
-    schema: Schema.Struct({ id: Schema.String, text: Schema.String, completed: Schema.Boolean }),
-  }),
-  todoUpdated: Events.synced({
-    name: 'todoUpdated',
-    schema: Schema.Struct({
-      id: Schema.String,
-      text: Schema.String.pipe(Schema.optional),
-      completed: Schema.Boolean.pipe(Schema.optional),
-    }),
-  }),
-  AppRouterSet: AppRouterSchema.set,
-  UserInfoSet: userInfo.set,
-  KvSet: kv.set,
-}
-
-const materializers = State.SQLite.materializers(events, {
-  todoCreated: ({ id, text, completed }) => todos.insert({ id, text, completed }),
-  todoUpdated: ({ id, text, completed }) => todos.update({ ...omitUndefineds({ completed, text }) }).where({ id }),
-})
-
-export const tables = { todos, app, userInfo, AppRouterSchema, kv }
-
-const state = State.SQLite.makeState({ tables, materializers })
-export const schema = makeSchema({ state, events })
-
-export const makeTodoMvcSolid = ({
-  otelTracer,
-  otelContext,
-}: {
-  otelTracer?: otel.Tracer | undefined
-  otelContext?: otel.Context | undefined
-} = {}): Effect.Effect<
+export const makeTodoMvcSolid = (
+  opts: CreateTodoMvcStoreOptions = {},
+): Effect.Effect<
   {
-    wrapper: ({ children }: any) => JSX.Element
-    store: Store<LiveStoreSchema<SqliteDsl.DbSchema, State.SQLite.EventDefRecord>, {}> & LiveStoreSolid.SolidApi
+    wrapper: ({ children }: any) => Solid.JSX.Element
+    store: Store<typeof schema> & LiveStoreSolid.SolidApi
   },
   UnknownError,
   Scope.Scope
 > =>
   Effect.gen(function* () {
-    const store: Store<any> = yield* createStore({
-      schema,
-      storeId: 'default',
-      adapter: makeInMemoryAdapter(),
-      debug: { instanceId: 'test' },
-    })
+    const store = yield* createTodoMvcStore(opts)
 
     const storeWithSolidApi = LiveStoreSolid.withSolidApi(store)
 
-    const wrapper = ({ children }: any) => <>{children}</>
+    const wrapper = (props: Solid.ParentProps) => <>{props.children}</>
 
     return { wrapper, store: storeWithSolidApi }
-  }).pipe(provideOtel(omitUndefineds({ parentSpanContext: otelContext, otelTracer })))
+  })
