+++ packages/@livestore/solid/tsconfig.json
@@ -15,6 +15,7 @@
     { "path": "../adapter-web" },
     { "path": "../livestore" },
     { "path": "../utils" },
-    { "path": "../utils-dev" }
+    { "path": "../utils-dev" },
+    { "path": "../../../tests/package-common" }
   ]
 }
