+++ tests/package-common/src/todomvc-fixture.ts
@@ -1,0 +1,137 @@
+import { makeInMemoryAdapter } from '@livestore/adapter-web'
+import { provideOtel, type UnknownError } from '@livestore/common'
+import { Events, makeSchema, State } from '@livestore/common/schema'
+import type { Store } from '@livestore/livestore'
+import { createStore } from '@livestore/livestore'
+import { omitUndefineds } from '@livestore/utils'
+import { Effect, Schema, type Scope } from '@livestore/utils/effect'
+import type * as otel from '@opentelemetry/api'
+
+// ============================================================================
+// Types
+// ============================================================================
+
+export type Todo = {
+  id: string
+  text: string
+  completed: boolean
+}
+
+export type Filter = 'all' | 'active' | 'completed'
+
+export type AppState = {
+  newTodoText: string
+  filter: Filter
+}
+
+// ============================================================================
+// Tables
+// ============================================================================
+
+const todos = State.SQLite.table({
+  name: 'todos',
+  columns: {
+    id: State.SQLite.text({ primaryKey: true }),
+    text: State.SQLite.text({ default: '', nullable: false }),
+    completed: State.SQLite.boolean({ default: false, nullable: false }),
+  },
+})
+
+const app = State.SQLite.table({
+  name: 'app',
+  columns: {
+    id: State.SQLite.text({ primaryKey: true, default: 'static' }),
+    newTodoText: State.SQLite.text({ default: '', nullable: true }),
+    filter: State.SQLite.text({ default: 'all', nullable: false }),
+  },
+})
+
+const userInfo = State.SQLite.clientDocument({
+  name: 'UserInfo',
+  schema: Schema.Struct({
+    username: Schema.String,
+    text: Schema.String,
+  }),
+  default: { value: { username: '', text: '' } },
+})
+
+const AppRouterSchema = State.SQLite.clientDocument({
+  name: 'AppRouter',
+  schema: Schema.Struct({
+    currentTaskId: Schema.String.pipe(Schema.NullOr),
+  }),
+  default: {
+    value: { currentTaskId: null },
+    id: 'singleton',
+  },
+})
+
+const kv = State.SQLite.clientDocument({
+  name: 'Kv',
+  schema: Schema.Any,
+  default: { value: null },
+})
+
+// ============================================================================
+// Events
+// ============================================================================
+
+export const events = {
+  todoCreated: Events.synced({
+    name: 'todoCreated',
+    schema: Schema.Struct({ id: Schema.String, text: Schema.String, completed: Schema.Boolean }),
+  }),
+  todoUpdated: Events.synced({
+    name: 'todoUpdated',
+    schema: Schema.Struct({
+      id: Schema.String,
+      text: Schema.String.pipe(Schema.optional),
+      completed: Schema.Boolean.pipe(Schema.optional),
+    }),
+  }),
+  AppRouterSet: AppRouterSchema.set,
+  UserInfoSet: userInfo.set,
+  KvSet: kv.set,
+}
+
+// ============================================================================
+// Materializers & Schema
+// ============================================================================
+
+const materializers = State.SQLite.materializers(events, {
+  todoCreated: ({ id, text, completed }) => todos.insert({ id, text, completed }),
+  todoUpdated: ({ id, text, completed }) => todos.update({ ...omitUndefineds({ completed, text }) }).where({ id }),
+})
+
+export const tables = { todos, app, userInfo, AppRouterSchema, kv }
+
+const state = State.SQLite.makeState({ tables, materializers })
+export const schema = makeSchema({ state, events })
+
+// ============================================================================
+// Store creation helper
+// ============================================================================
+
+export type CreateTodoMvcStoreOptions = {
+  otelTracer?: otel.Tracer | undefined
+  otelContext?: otel.Context | undefined
+}
+
+/**
+ * Creates a TodoMVC store for testing. Framework-specific wrappers should
+ * use this function and add their own API bindings.
+ */
+export const createTodoMvcStore = ({
+  otelTracer,
+  otelContext,
+}: CreateTodoMvcStoreOptions = {}): Effect.Effect<Store<typeof schema>, UnknownError, Scope.Scope> =>
+  Effect.gen(function* () {
+    const store: Store<typeof schema> = yield* createStore({
+      schema,
+      storeId: 'default',
+      adapter: makeInMemoryAdapter(),
+      debug: { instanceId: 'test' },
+    })
+
+    return store
+  }).pipe(provideOtel(omitUndefineds({ parentSpanContext: otelContext, otelTracer })))
