+++ tests/package-common/tsconfig.json
@@ -13,6 +13,7 @@
     { "path": "../../packages/@livestore/utils-dev" },
     { "path": "../../packages/@livestore/sqlite-wasm" },
     { "path": "../../packages/@livestore/adapter-node" },
+    { "path": "../../packages/@livestore/adapter-web" },
     { "path": "../../packages/@livestore/livestore" }
   ]
 }
